+++ tests/protocols/client_server_handshake.rs
@@ -1,4 +1,5 @@
 use besedarium::*;
 
 // Client-server handshake (HTTP request/response)
-pub type HttpHandshake = TInteract<Http, TClient, Message, TInteract<Http, TServer, Response, TEnd<Http>>>;
+pub type HttpHandshake =
+    TInteract<Http, TClient, Message, TInteract<Http, TServer, Response, TEnd<Http>>>;