+++ tests/protocols/pubsub.rs
@@ -1,7 +1,8 @@
 use besedarium::*;
 
 // Publish/subscribe (MQTT)
-pub type MqttPubSub = TChoice<Mqtt,
+pub type MqttPubSub = TChoice<
+    Mqtt,
     TInteract<Mqtt, TClient, Publish, TEnd<Mqtt>>,
-    TInteract<Mqtt, TClient, Subscribe, TEnd<Mqtt>>
+    TInteract<Mqtt, TClient, Subscribe, TEnd<Mqtt>>,
 >;